requirements.txt
Django==1.4.3
Markdown==2.2.1
Pygments==1.6
South==0.7.6
argparse==1.2.1
django-markdown-deux==1.0.3
django-openid-auth==0.4
django-ratings==0.3.7
django-tagging==0.3.1
markdown2==2.1.0
python-openid==2.2.5
wsgiref==0.1.2
